Health bars
Created a stats blueprint component for tracking stats
Created an enumerator for stats
Added a function to the component pointing at the enumerator
Added the component to 3rd person bp and edited mesh to collision enabled query and physics
Added damage value to BPC_Combat
Damage is applied when the player hits the target. It is, however, applied every .3 seconds not based on the animation. That. Is. Bad.
Created a hit react montage based on the added animation.
Created a player health widget, added a canvas and progress bar,
Created a HUD wb and added the player health widget to it
- when adding player health's anchor, held left control and set it to full
Set up health bar to lower based on damage
Created an enemy health bar widget - not hooked it up yet