AI foundation
Reduced camera lag on boom of 3rdpersoncharacter from 8 to 5
Created a blackboard, Behavior Tree, and AIControllerBP for the AI
Made a new enemy bp, attached the AIControllerBP to it via Class Defaults
Created an AI task for Chasing
Added a navmesh to the level
Set Enemy BP Mesh and Capsule to ignore the camera collision
Added pawn sensing to the BP AI Component with all hearing at 0
Added a decorator to the blackboard, setting the observer to self, key query to not set, and blackboard key to seen target. Decorators are conditions. They are called decorators because conditions would be a more appropriate terms clearly identifying their functionality. Can't have clarity.