Added boss foundation
Created Boss BP, Behavior Tree, added that to a boss AI BP (all children).
Added boss slam animation and associated that with a new sequence in the Boss BT.
Created anim notify bp for doing a point trace for applying damage
Added that to the damage bp interface
Added the notify to the boss attack animation montage when it hits
Created an arrow on the BP Enemy for the boss attack's point trace position
Altered the Acceptance Radius of Task_ChaseTarget to account for player location - in the Boss BT, changed that to a larger number.
Set up boss big hit attack to deal damage