Wave Spawner
It almost works flawlessly.
The waves spawn if a bool on the spawner is set to active. The enemy waves begin, spawning up to a max number of enemies set by the dev. When enemies die they run a function on the spawner to respawn another wave IF all the enemies of that wave are dead. If all enemies are dead and the number of waves are over, a print string appears saying YOU WIN PLUS INFINITY!!!